variable type fixes and msvc tweaks
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@1262 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_core.c
@@ -2393,7 +2393,7 @@ static void *SWITCH_THREAD_FUNC switch_core_sql_thread(switch_thread *thread, vo
 		}
 		
 		if (diff < freq) {
-			diff = (switch_time_now() - last_commit) / 1000;
+			diff = (uint32_t)((switch_time_now() - last_commit) / 1000);
 		} 
 
 		if (trans && (itterations == target || diff >= freq)) {
